moved gallery card into a component, set up for likes fix
