+++ src/components/Gallery.js
@@ -1,39 +1,12 @@
-import { FaRegTrashAlt } from 'react-icons/fa';
 import firebase from '../firebase';
-import { getDatabase, ref, onValue, remove, update } from 'firebase/database';
+import { getDatabase, ref, onValue } from 'firebase/database';
 import { useEffect, useState } from 'react';
-import { useUserAuth } from '../context/UserAuthContext.js';
-import DeleteConfirmation from './DeleteConfirmation.js';
+import GalleryCard from './GalleryCard';
 
 function Gallery({ closeGallery, showGallery }) {
   //useState for gallery of user's backronym
   const [gallery, setGallery] = useState([]);
   const [backronymFilter, setBackronymFilter] = useState('recent');
-  const [deleteWarning, setDeleteWarning] = useState(false);
-  const [deleteID, setDeleteID] = useState('');
-  let { user } = useUserAuth();
-
-  // delete entry
-  function handleDelete(e, resultKey) {
-    e.preventDefault();
-    const database = getDatabase(firebase);
-    const dbRef = ref(database, `/${resultKey}`);
-    remove(dbRef);
-
-    //reset delete ID state
-    setDeleteID('');
-  }
-
-  //updating like count
-  function handleLike(resultKey, resultLikes) {
-    const updatedLikes = {
-      likes: resultLikes + 1,
-    };
-
-    const database = getDatabase(firebase);
-    const childRef = ref(database, `/${resultKey}`);
-    update(childRef, updatedLikes);
-  }
 
   //connect to firebase when Gallery component mounts
   useEffect(() => {
@@ -111,66 +84,7 @@ function Gallery({ closeGallery, showGallery }) {
             {
               // map over the gallery state (from firebase). Results is each submission
               gallery.map((result) => {
-                return (
-                  <li className="galleryCard" key={result.key}>
-                    <h3>{result.userInput}</h3>
-                    {/* mapping over each user's submission results array item (each word in array is the initial) */}
-                    {result.results.map((initialWord, index) => {
-                      return <p key={`${result.key}-${index}`}>{initialWord}</p>;
-                    })}
-
-                    <div className="userGalleryControls">
-                      {user === null ? (
-                        ''
-                      ) : user.email === result.email ? (
-                        <button
-                          className="deleteBtn"
-                          onClick={(e) => {
-                            setDeleteWarning(true);
-                            setDeleteID(result.key);
-                          }}
-                        >
-                          <span className="sr-only">Delete</span>
-                          <FaRegTrashAlt />
-                        </button>
-                      ) : user.email === null && result.email === 'anonymous' ? (
-                        <button
-                          className="deleteBtn"
-                          onClick={(e) => {
-                            setDeleteWarning(true);
-                            setDeleteID(result.key);
-                          }}
-                        >
-                          <span className="sr-only">Delete</span>
-                          <FaRegTrashAlt />
-                        </button>
-                      ) : (
-                        ''
-                      )}
-
-                      {deleteWarning ? (
-                        <DeleteConfirmation
-                          setDeleteWarning={setDeleteWarning}
-                          handleDelete={handleDelete}
-                          deleteID={deleteID}
-                        />
-                      ) : null}
-
-                      {user ? (
-                        <button
-                          className="likeBtn"
-                          onClick={() => {
-                            handleLike(result.key, result.likes);
-                          }}
-                        >
-                          <span className="sr-only">Like</span>
-                        </button>
-                      ) : null}
-
-                      <p className="likeCount">{result.likes}</p>
-                    </div>
-                  </li>
-                );
+                return <GalleryCard result={result} key={result.key} />;
               })
             }
           </ul>
